Fix lint errors in client
