+++ client/src/components/Tables/DynamicTable/TableRow.tsx
