+++ client/src/components/Tables/DynamicTable/styles.scss
